fixup(food): PRD-150 Copilot R1 — 8 inline findings addressed
Addresses every Copilot finding on PR #2824:

1-3. Replaced three `<Link><Button/></Link>` nestings with `<Button asChild><Link/></Button>` so each CTA renders as a single interactive `<a>` instead of `<a><button/></a>`. Fixes: FridgePage "What can I cook?" header link, SolvePage empty-state Open-fridge link, SolveRecipeCard "Cook this" link.

4. `TagsInput.commit` now de-duplicates entries before passing them to the solver — the previous parser preserved "vegan, vegan" and the server-side `COUNT(DISTINCT tag) = required.length` AND-overlap check would then never match.

5. `TagsInput` is still uncontrolled (matches the project's tag-input convention) but now keys the `<input>` on the current `filters.tags` join, forcing React to remount it whenever a reset / external state change moves the canonical value — stale `defaultValue` can no longer shadow a "Clear filters" reset.

6. `parseContextTags` now catches `JSON.parse` failures and treats malformed rows as `[]` (wildcard). A single corrupt `substitutions.context_tags` row no longer crashes the whole solver request.

7. `loadSubstitutionsIndex(db, recipeIds)` now applies the scope filter in SQL: WHERE `scope='global'` OR (`scope='recipe'` AND `recipe_id IN (...)`). When callers pass a recipe set, the query only loads the edges that can possibly apply.

8. `recipe_lines.qty_g/qty_ml/qty_count` of NULL now flows through as `qty: null` and the line-evaluator fails closed for that line — a recipe with any unresolved canonical qty is never declared cookable. New regression test pins the behaviour.

File-size cleanup: the new SQL helpers pushed `substitutions-resolve.ts` past the 200-line oxlint cap. Extracted SQL loaders to `substitutions-resolve-loaders.ts` and shared types to `substitutions-resolve-types.ts`; the public service file re-exports the same surface so no caller needs to change its import path.

Tests:
- New "fails closed when a required line has an unresolved canonical qty" case in `solver.test.ts` (24 cases total across the two solver files).
- Existing 23 server cases + 6 RTL cases still green.
